Extracted the code into a separate module.
